+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ### 0.1.0
+- Add option to choose draw mode in [DrawItem.draw()](https://github.com/noah-friedman/obj.GL/blob/v0.1.0/src/DrawItem/index.ts#L131)
 
 ### 0.0.6
 - Add [.npmignore](https://github.com/noah-friedman/obj.GL/blob/v0.0.6/.npmignore)